refactor: daily recommendation 파일명 정리 #65
