+++ fe/src/services/daily-recommendation/base.ts
